+++ src/components/pod/pod.component.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useMemo } from "react";
 import { MarginProps } from "styled-system";
-import Logger from "../../__internal__/utils/logger";
 import useLocale from "../../hooks/__internal__/useLocale";
 
 import {
@@ -45,12 +44,9 @@ export interface PodProps extends MarginProps {
   /** A component to render as a Pod footer */
   footer?: string | React.ReactNode;
   /** Supplies an edit action to the pod */
-  onEdit?:
-    | string
-    | Record<string, unknown>
-    | ((
-        ev: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>
-      ) => void);
+  onEdit?: (
+    ev: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>
+  ) => void;
   /** Supplies a delete action to the pod */
   onDelete?: (
     ev: React.MouseEvent<HTMLElement> | React.KeyboardEvent<HTMLElement>
@@ -72,9 +68,6 @@ export interface PodProps extends MarginProps {
   /** Renders edit button inside the pod if it exists. */
   internalEditButton?: boolean;
 }
-
-let deprecationOnEditStringWarnTriggered = false;
-let deprecationOnEditObjectWarnTriggered = false;
 
 const Pod = React.forwardRef<HTMLDivElement, PodProps>(
   (
@@ -103,20 +96,6 @@ const Pod = React.forwardRef<HTMLDivElement, PodProps>(
     }: PodProps,
     ref
   ) => {
-    if (!deprecationOnEditStringWarnTriggered && typeof onEdit === "string") {
-      deprecationOnEditStringWarnTriggered = true;
-      Logger.deprecate(
-        "Support for passing strings to the `onEdit` prop of the `Pod` component is now deprecated. Please only pass event handlers to `onEdit`."
-      );
-    }
-
-    if (!deprecationOnEditObjectWarnTriggered && typeof onEdit === "object") {
-      deprecationOnEditObjectWarnTriggered = true;
-      Logger.deprecate(
-        "Support for passing objects to the `onEdit` prop of the `Pod` component is now deprecated. Please only pass event handlers to `onEdit`."
-      );
-    }
-
     const [isEditFocused, setEditFocused] = useState(false);
     const [isEditHovered, setEditHovered] = useState(false);
     const [isDeleteFocused, setDeleteFocused] = useState(false);
@@ -158,10 +137,8 @@ const Pod = React.forwardRef<HTMLDivElement, PodProps>(
       onMouseLeave: () => setEditHovered(false),
       onFocus: () => setEditFocused(true),
       onBlur: () => setEditFocused(false),
-      ...(typeof onEdit === "function" && {
-        onClick: processPodAction(onEdit),
-        onKeyDown: processPodAction(onEdit),
-      }),
+      onClick: onEdit && processPodAction(onEdit),
+      onKeyDown: onEdit && processPodAction(onEdit),
     };
 
     return (
@@ -264,7 +241,6 @@ const Pod = React.forwardRef<HTMLDivElement, PodProps>(
                 noBorder={!border}
                 size={size}
                 variant={variant}
-                {...(typeof onEdit === "string" ? { href: onEdit } : onEdit)}
               >
                 <Icon type="edit" />
               </StyledEditAction>