+++ src/style/utils/filter-styled-system-padding-props.ts
@@ -1,21 +1,48 @@
 import { PaddingProps } from "styled-system";
 import filterObjectProperties from "../../__internal__/filter-object-properties";
 
-export const paddingPropertyNames = [
-  "padding",
-  "p",
-  "paddingLeft",
+export const paddingNames: (keyof PaddingProps)[] = ["p", "padding"];
+
+export const paddingLeftPropertyNames: (keyof PaddingProps)[] = [
   "pl",
-  "paddingRight",
+  "paddingLeft",
+];
+
+export const paddingRightPropertyNames: (keyof PaddingProps)[] = [
   "pr",
-  "paddingTop",
+  "paddingRight",
+];
+
+export const paddingXPropertyNames: (keyof PaddingProps)[] = ["px", "paddingX"];
+
+const paddingHorizontalPropertyNames = [
+  ...paddingLeftPropertyNames,
+  ...paddingRightPropertyNames,
+  ...paddingXPropertyNames,
+];
+
+export const paddingTopPropertyNames: (keyof PaddingProps)[] = [
   "pt",
-  "paddingBottom",
+  "paddingTop",
+];
+
+export const paddingBottomPropertyNames: (keyof PaddingProps)[] = [
   "pb",
-  "paddingX",
-  "px",
-  "paddingY",
-  "py",
+  "paddingBottom",
+];
+
+export const paddingYPropertyNames: (keyof PaddingProps)[] = ["py", "paddingY"];
+
+const paddingVerticalPropertyNames = [
+  ...paddingTopPropertyNames,
+  ...paddingBottomPropertyNames,
+  ...paddingYPropertyNames,
+];
+
+export const paddingPropertyNames = [
+  ...paddingNames,
+  ...paddingHorizontalPropertyNames,
+  ...paddingVerticalPropertyNames,
 ];
 
 export default function filterStyledSystemPaddingProps(